Add some notes on common behaviours of clients and servers.
A lot of the code in the Client module could be moved into a common
'EndPoint' module, and then be specialized for building a Client and
a Server. I've jotted down a few notes on what I think is needed to
allow extracting the common code.

The most pressing thing I think is to change the frame types so we
can deal with different kinds of frames as different types. This
can make it easier to plug in responses for different kinds of frames
that must be treated a bit differently by clients and servers, and
also make it impossible to accidentally use the wrong field selector
with the wrong variant of a multi-variant type.

That rework isn't difficult, but it will be messy, so I will perform
it in several small commits, starting next.