Move the DataFrame fields into a Data type.
This is the first step in a large reworking of the frame types, aimed
at allowing client and server code to treat individual frames in a
more abstract way. It will also make it impossible to use a field
selector for one variant of frame with another.

The unit tests still pass, and the test programs still build and work.