Rename many frame data constructors, and the ControlFrameDetails type.
I was considering putting a SPDYVersion field inside each of the control
frame types SynStream through Credential, removing the 'spdyVersion'
field from the ControlFrame variant of Frame, and then flattening the
ControlFrameDetails type into Frame. However, I tried this and got
to the point of flattening ControlFrameDetails into Frame when I
realized it would create some unfortunate and tedious code duplication
in the Internal.Serialize module. So I've decided to keep the data
structures 'normalized', with the SPDY version in just one place.

However, I've decided to do a bunch of renaming, in particular:

- the data constructors for Frame are now called 'AControlFrame'
  and 'ADataFrame' instead of 'ControlFrame' and 'DataFrame'

- the ControlFrameDetails type is now called ControlFrame, and each
  of the data constructors of that type are now named 'A<X>Frame'
  for each different type of control frame

- the control frame content types SynStream through Credential are
  now called SynStreamFrame, ..., CredentialFrame. This, despite the
  fact that they do not themselves carry the SPDY version

- I've also removed the field selectors for the AControlFrame
  variant of the Frame type

Since the change in name for from Data to DataFrame would also change
the indentation of that type's field selectors and API docs, I decided
to reformat the declaration in the same way that the other frame types
are formatted. Other than renaming the type, there are no functional
differences in the code though.

I also changed the word 'details' in some pattern matches against
ControlFrameDetails values to 'cf' for 'control frame'.

Aside from the renamings and the required compensating changes in other
modules, there are no functional changes in this code.