Add the FrameHandlers type and accompanying 'handleFrame' function.
In order to abstract out the handling of frames in the Client module
so that we can share code with an upcoming Server module, I've added
a FrameHandlers type and a 'handleFrame' function that takes care
of applying the right handler to the right frame type. The handler
for each control frame type takes the SPDY version in a separate
argument.

In the next commit, I'll rework the code for the frame reader thread
in the Client module so that it is parametrized by a FrameHandler.