Rework Client.readFrames to use FrameHandlers.
The 'handleFrame' function in 'readFrames' now uses Frames.handleFrame
with a set of client-specific frame handlers. This gets us part of
the way to making 'readFrames' generic for any kind of SPDY endpoint.

Of course, 'readFrames' still invokes the 'clientInputFrameHandlers'
function, so we need to enhance a 'Connection' with a field that
carries a function for producing the input handlers. I'll turn that
into a separate commit.

In the process, I've added a 'defaultIOFrameHandlers' value to
the Frames module. This is convenient since there are still several
handlers that need to be written.

At this point, the 'sping' and 'sget' programs still compile and
work properly.