Pull the frame handlers out of Client.readFrames.
Now the 'readFrames' function takes a set of frame handlers, instead
of calling 'clientInputFrameHandlers' itself on the
connection. Instead, 'setupConnection' creates the handlers.

That's not really where we want 'clientInputFrameHandlers' called
either, because there is a lot about 'setupConnection' that could be
common to either a client or a server endpoint. But 'setupConnection'
already has the client-specific code for connecting to a server, so
this is an OK resting point for client-specific code for the
moment. The next task will be to extract the endpoint-agnostic code
from this function.