Put the function for creating input frame handlers in Client.
Rather than directly calling 'clientInputFrameHandlers' in the
'setupConnection' function, add a field to the Client type that
holds the function for creating input frame handlers.

At some point, Client will probably be generalized to an EndPoint
type, and contain all endpoint-varying functionality.

In the process, I've renamed the original 'clientInputFrameHandlers'
function to 'stdInputFrameHandlers'.

The test programs still compile and work with the 'flip' server.