Move creating the network connection to 'getConnection'.
Instead of establishing the network connection in 'setupConnection',
establish it in 'getConnection'. This way, 'setupConnection' becomes
more agnostic to the type of endpoint. The 'getConnection' function
can remain client-specific.

The code change here is very small, but because of removing the
'let' in 'setupConnection' the indentation of the whole function
changes. Almost all the changes here are reductions in the leading
whitespace in that function.

When we switch to supporting TLS, we'll need to use something
different from a raw IO handle for representing the underlying
network connection. When using hs-tls, we'll instead have a TLSCtx
container around the handle. I think it would be nice to be able
to write a SPDY server that does not insist on TLS, so we'll want
to create some kind of type for abstracting over the network channel,
one that knows how to send bytes, receive them, flush, and close.
I'll have to think a little bit about whether I need to do that before
or after splitting out a generic EndPoint module.