Add a sample client and server demonstrating the use of hs-tls.
The simplest way to use TLS is to use the hs-tls library, particularly
since there is an experimental branch that implements the NPN
extension. Before trying out the NPN extension though, I decided to
first figure out the basics of using TLS.

I've added

 tls-server

   a sample TLS server that sends the string 'hello\n' to each
   client that connects. You need to supply a certificate and
   a private key in 'pem' format, the default format produced
   by the openssl command line tool

 tls-client

   a matching client that connects to the server, receives the
   string, and prints it

 notes/tls-ins-and-outs.txt

   a set of notes on getting tls-server and tls-client working.

The make file and .gitignore are updated for the tls-server and
tls-client programs.

I've used John Millikin's 'options' package to give the two
test programs some command-line options, though I ordinarily
just use the baked-in defaults.

At this point, tls-client can connect to tls-server and correctly
receive the string. The next step is to try out the NPN extension
by advertising a list of protocols on the server, and selecting
one on the client. I figure I'll add a command-line option to
the client that gives the desired protocol string. On the server
side, the different protocols will just send different messages
to the connected client.