Use the NPN extension in the tls-server and tls-client demo programs.
Using the NPN extension is pretty straightforward:

- on the server side, define the 'onSuggestNextProtocols' field
  in the TLSParams, so it provides a list of supported protocols

- on the server side, use the 'getNegotiatedProtocol' function
  to determine the protocol, and take action depending on which
  one was chosen

- on the client, define the 'onNPNServerSuggest' field in the
  TLSParams, so it picks a protocol.

- on the client, use the 'getNegotiatedProtocol' function to
  determine the protocol, and take action depending on which one
  was chosen.

Here, the server implements two protocols, 'hello' and 'time'. For
'hello', it sends the string 'hello\n' to the client, and for 'time'
it sends the server's local time of day.

The client ignores the protocols suggested by the server, and
always selects either 'hello', or the protocol name selected
by the user.

There is no automatic rejection of a protocol that does not appear
in the server's suggested protocols, so on the server side we
must handle an unknown protocol. Apparently, the hs-tls implementation
will pass through the client's requested protocol even if it does
not appear in the list of suggestions. I've chosen to simply send
an 'unknown protocol' message so it's clear that the server has
received the bogus protocol. In practice, I imagine it would be
equally correct to simply close down the TLS connection, though
for SPDY it's probably a good idea to send a GOAWAY frame first.

Getting this working was relatively painless compared to getting
TLS set up at all. Now we're in a position to integrate TLS into
the SPDY client, and the 'sping' and 'sget' test programs. We
can then try out hs-tls against the TLS implementation used by
the 'flip' server.