Add a NetworkConnection module, and abstract over Handles in Client.
In order to support TLS connections and non-TLS connections, we
cannot directly use socket handles in the Client. Instead, define
a NetworkConnection as a collection of four functions,

 sendData

   for sending bytes

 receiveData

   for receiving bytes

 flush

   for forcing buffered bytes out onto the wire

 close

   for closing the connection

There are functions for turning a Handle or a (TLSCtx Handle)
into a NetworkConnection.

In Client, I've replaced Connection.connSocketHandle with
Connection.connNetworkConnection, and used 'fromHandle' in the
'getConnection' function.

The 'sping' and 'sget' functions still work with this change, though
I had to change the GHC options in the Makefile to exclude the
the 'monad-tf' package, since there is a conflict with 'mtl'.
Unfortunately, this breaks liking 'sget-prof'. I'll have to look
into optionally compiling the test programs using Cabal instead
of make; I'm not terribly interested in figuring out how to give
GHC the correct options here.

That can wait for a few commits though. The more pressing task is
to enable TLS connections in the client, and see whether we can
talk to a 'flip' server when it's using TLS.