Make TLS connections the default in the Client.
With this change, and a minor change in the 'flip' server, we can
now talk to 'flip' when it's using TLS.

A client now has an option for a ConnectionStyle. There are three
types of connection styles:

 CsTLS

   the default style of connection, which uses TLS on a socket
   handle

 CsSocket

   just use a socket handle without TLS. Up until now, this was
   the only option we offered

 CsCustom

   supply a function that takes a ConnectionKey and establishes
   a low-level connection to the remote endpoint. This is intended
   for internal testing

I've made CsTLS the default ConnectionStyle.

The second part of the change is the 'toNetworkConnection' function,
which takes a connection style and a connection key, and attempts
to establish a connection. The implementation of the case for
CsTLS is taken almost verbatim from the 'tls-client' example program.

This stuff still has several rough edges:

- our approach to negotiating the protocol is to print the offered
  protocols on stderr and then pick 'spdy/3', whether it is among
  the offered protocols or not. We should introduce a proper logger
  and at least warn if the offered protocols do not include SPDY 3.

  In fact, since it's possible that the server will only offer
  HTTP, or will offer nothing -- in which case the standard approach
  would be to assume we're talking to an HTTP server -- you could
  argue that we shouldn't be establishing connections here in the
  client. Either that, or we should support an interface for
  supplying an HTTP client to handle HTTP. I'll think about that
  a little later.

- lousy error handling. In particular, we don't ensure that the
  socket handle is closed if there is an error in the NPN negotiation
  or a handshake failure or any other kind of IO error.

- lousy error representations when there is no negotiated protocol
  or when the negotiated protocol isn't what we expected.

With respect to the third point, it appears that the client has
the final say on the negotiated protocol, at least as far as hs-tls
is concerned. However, the interface suggests that the negotiation
could fail, so I've tried to program so as to detect a failure in
the negotiation. The problem is in using 'userError' to form the
IOErrors that we're throwing. It would be better to have some kind
of exception specific to these error cases.

On the 'flip' server side, I changed to it advertise 'spdy/3' as
an available protocol, and to accept 'spdy/3' as a valid protocol
selection from the client. Without that last change, the server
would throw an exception and die.

With these changes, rough as they are, both the 'sping' and 'sget'
sample programs run correctly against the 'flip' server when it's
using TLS. This was all relatively painless.

The changes required adding the 'crypto-api' and 'tls-extra'
packages to the dependencies in the cabal file.