Fix bug: the 'cftCredential' constant should be 0xA, not 0xB.
Tatsuhiro Tsujikawa asked a question in spdy-dev on April 3 about
whether the frame type code for CREDENTIAL should be 10 (0xA) or
11 (0xB). It was given as 11 in the spec, but apparently in the
Chromium source code it was 10. William Chan said that it is in fact
an error in the spec.

I thought I'd just fix this now before I forget. The unit tests still
pass, and this has no bearing on the client implementation as yet
since we don't use CREDENTIAL frames.