Add a note on how to create common EndPoint functionality.
I'm not entirely sure whether EndPoint should be a type or a
type class. Since there will almost certainly be only two
concrete types of endpoints (clients and servers), I think
that a type class is overkill. Best thing to do is to try it
out and see whether I'm wrong.