Create an Endpoint type from the 'ep' fields in Client.
Now Client has an 'endpoint' field of type Endpoint, and in the
few places where the 'ep' fields were used I've inserted calls to
'endpoint'. The next step is to figure out whether the functions
that use the 'ep' fields should still operate on a Client, or
whether they should be changed to operate on an Endpoint.