Use an Endpoint in a Connection, prepare to split the Client module.
Instead of a 'connClient' field, Connection now has a 'connEndpoint'
field. There are a number of modest changes in association with this,
intended to make it easy to separate Endpoint into its own module:

1. The 'setupConnection' function now takes an Endpoint instead of a
   Client

2. There is a new function 'getOrCreateConnection' which should allow
   us to hide the 'epConnectionMapMVar' field in the Endpoint type.

3. There is a new 'mkEndpoint' function which hides the details of
   creating a new endpoint. This also will help hide the EndPoint
   fields. I will likely rename this to 'endpoint' in a future commit.

At this point, I believe that most of the code in the Client module
can move into an EndPoint module. I'll make that change in the next
commit.