Split most of the Client module into a new Endpoint module.
This splits the code that deals only with Endpoints out into its
own module. I've preserved the names of functions and exposed just
the parts of the Endpoint module that are needed to compile Client.

This still has a few rough edges. In particular:

- a couple of functions (mkEndpoint, updateWindow') should be
  renamed

- the interface for performing PINGs is too low-level. Some part
  of the Client.ping function needs to be extracted and turned
  into a higher-level function in Endpoint

I'll resolve those problems in the next couple of commits.