Introduce EndpointOptions, complete with the first ping ID.
Thinking about how to make a higher-level interface for pings,
I realized that all Connections were being given an initial ping ID
of 1, which is correct only for clients. In order to support both
clients and servers, we need to make the initial ping ID a parameter
of the endpoint, since clients must use odd ping IDs and servers
must use even ping IDs. In fact, we'll need to do something similar
for stream IDs, but I'll wait for a subsequent commit.

Since the 'endpoint' function now has to take more than one
configuration value, it makes sense to add an EndpointOptions type
and collect the frame handlers function and the initial ping ID
there. I've done this, changed the signature of 'endpoint', and
changed 'setupConnection' to get the initial ping ID from the
Endpoint.