Move most of the high-level ping facility to EndPoint.
All but one line of Client.ping can be moved to the Endpoint
module, if we extract a function 'pingRemote' that takes PingOptions
and a Connection.

The only remaining problem is that we call 'removePingHandler'
in the standard frame handler for PING frames defined in Client.
However, we could define this one handler function as a utility in
Endpoint, and simply use it in the Client module. I'll do that in
the next commit; once that's done, we should be able to hide the
'installPingHandler' and 'removePingHandler' functions in the
Endpoint module.