Handle ping frames in a default set of Endpoint frame handlers.
Instead of handling incoming ping frames in
Client.stdClientInputFrameHandlers, handle them in a default set
of handlers defined in the Endpoint module. This slightly simplifies
Client. With the new 'isInitiatedHere' function, we also get
client-side and server-side responses for pings in a single handler.

At this point, we should be able to hide 'installPingHandler' and
'removePingHandler' in the Endpoint module. I'll do that in the
next commit.