Add Endpoint.fromSockAddr and a new UnixSockKey variant of ConnectionKey.
We'll need a way to convert a socket address into a connection key
in order to support servers. In fact, arguably 'toConnectParams'
should probably produce a SockAddr, even if it needs to return a
result in the IO monad because of name lookups. This would allow
us to also support connections through named unix sockets.

In preparation for that, and to avoid throwing an error in
'fromSockAddr', I've added a UnixSockKey variant to ConnectionKey.
Strictly speaking, SPDY does not mention this kind of identifier
for a connection, but it doesn't prohibit it.