Add the Server module and the SServe test program.
Now we have a second kind of endpoint, the Server. A Server is an
endpoint that receives new connections from some other process through
the 'addConnection' function. There are standard server loops for
accepting connections on a socket, one without TLS (runSocketServer)
and one with TLS (runTLSServer). In fact, a single Server can be
fed incoming connections by more than one concurrent process, so it's
possible to have a server that serves TLS on one port and non-TLS
on another.

This approach would allow us to write a server that supported both
HTTP and SPDY when using TLS.

At the moment, a Server can't do much except answer pings. The next
task is to give a server a pluggable way to translate the headers
from a SYN_STREAM frame into the headers and data for a response.

Along with the Server module, I've added the SServe test program,
which starts a server on a port, by default using TLS but optionally
without. The SPing program can successfully ping SServe with and
without TLS -- so we've got a decent start.

Finally, I've made the necessary additions in the cabal file and
the Makefile.