Another request handler revision: half-closed requests and responses.
The current type for a request handler forces the server to invoke
the resulting IO action in order to determine whether to set the
FIN flag on a SYN_REPLY frame. This is bad, because if the remote
end is waiting for the SYN_REPLY before sending DATA or HEADERS
frames, we'll live lock.

Instead, the handler optionally receives an IO action to get the
stream contents, and optionally gives one to get the response
contents. Invoking a request handler with the arguments

   headers Nothing

means that the SYN_STREAM frame had the FIN flag set, since there
is no way to get further content on the stream. Returning a value
of the form

   (headers, Nothing)

means that there is no further content on the response, so the
FIN flag can be set on the SYN_REPLY frame.