Notes on hiding flow control from the web app and client app layers.
I've come to the conclusion that it's best to hide flow control from
the upper layers, and have added some notes on why. I'm starting to
think that a hybrid push-pull model would simplify the upper layer:
pull for request data and headers, push for response data and headers.

The next step is to create a bounded buffer data structure that we
can use in streams.