Add the Internal.BoundedBuffer module.
In order to hide per-stream buffering and flow control inside an
endpoint, it will be useful to have a bounded buffer that blocks
the consumer when its empty, and blocks the consumer when it's full.

I couldn't find an implementation of the idea on hackage, which
could mean (a) it's a stupid idea, or (b) it's called something else,
or (c) there is already a slightly different thing that can do
what's needed, but I'm not smart enough to realize it. But the
concept is straightforward. Unit tests to come.