Generalize BoundedBuffer to hold 'sized' values.
At first I thought that it would be sufficient to bake ByteStrings
into BoundedBuffer, but I've changed my mind. In order to preserve
the relative order of headers and data chunks, it would be useful
to have the buffer hold some kind of stream content type, something
like

  Either HeaderBlock ByteString

For flow control, only the data chunks count, so we'd want a 'size'
function along the lines of

  size :: Either HeaderBlock ByteString -> Int
  size = either (const 0) B.length

where HeaderBlocks are zero-size and ByteStrings have size equal
to their length.

I've consequently generalized BoundedBuffer by adding a Sized
type class, and changing the BoundedBuffer type so it can hold
any type, though most of the functions require a Sized type.