Switch to using the semaphores from SafeSemaphore in BoundedBuffer.
It turns out that it's going to be important to have a non-blocking
version of 'add' for BoundedBuffer, one that returns True if a chunk
was added, and returns 'False', without blocking indefinitely, if it
couldn't be. When the remote endpoint does not properly implement
flow control, this allows the reader thread to signal an error
instead of blocking for some length of time, or spawning off a
thread to do the addition when space is available.

The problem is that QSemN does not provide a way to acquire a
quantity without blocking if it's not immediately available.

After browsing on Hackage for a bit, I noticed the SafeSemaphore
package by Chris Kuklewicz. This package offers replacements for
QSem and QSemN called MSem and MSemN, which are safer in the sense
that they handle the case where waiting threads throw exceptions.
However, they also have a somewhat richer interface, including
a way to acquire a quantity only when it's available.

This commit switches to MSem and MSemN. In the next commit, I'll
add a new function for the interface for adding to the buffer only
when there is enough space.