Add a 'tryAdd' function to BoundedBuffer.
The new function

  tryAdd :: Sized a => BoundedBuffer a -> a -> IO Bool

either adds a chunk and returns True, or returns False without
blocking indefinitely. The MSemN.waitF function is critical here,
since it does not block when we ask for a quantity of zero. This
lets us ask for zero bytes when there isn't enough space for the
chunk, and we can also return a Bool indicating whether there
was enough space for the chunk -- even when we ask for zero
because of a zero-sized chunk.

I've also added a unit test that confirms that 'tryAdd' does not
block when the chunk cannot be added, and that it otherwise behaves
as 'add'.

What's not right here is we should be using 'bracket' or a similar
utility to make sure that exceptions don't leak quantities from
the semaphore. That is true in more than one place, so I'll attempt
that in another commit.