Extract common code in BoundedBuffer.{add, tryAdd}.
It's possible for an asynchronous exception delivered during 'add',
'tryAdd', or 'remove' to corrupt the internal state of the buffer.
Before solving that problem, I've decided to create a function
'addGeneral' that optionally blocks until space is available, and
use it in 'add' and 'tryAdd'. The idea is to avoid repeating the
code for masking asynchronous exceptions.