Protect against asynchronous exceptions in BoundedBuffer.
If a thread receives an asynchronous exception in 'addGeneral' or
'remove', the buffer's internal state could be corrupted. For example,
if the exception arrives immediately after taking the required
quantity from the available space semaphore, we won't add the chunk
and will consequently leak the requested space.

After reading over the API docs for SafeSemaphore and
Control.Exception, it looks like it is sufficient to simply
use 'mask_' on a block of code that

 - waits on one semaphore
 - modifies the sequence of chunks
 - signals on the other semaphore

The SafeSemaphore wait functions are interruptible, but if they
are interrupted they don't lose any of the semaphore quantity.
Data.IORef.atomicModifyIORef is not interruptible, and neither
are the SafeSemaphore signal functions. Therefore, using 'mask_'
the only place where an asynchronous exception can be thrown
is in the wait. Since the semaphore will not lose any of its
quantity in this case, it's safe to simply let the exception
propagate. If the wait returns, then the rest of the computation
is not interruptible, and 'mask_' will delay the arrival of
asynchronous exceptions until we're done modifying the chunk
sequence and signalling the other semaphore. There is no need
to use 'uninterruptibleMask_' ourselves.