Interface change: give the client a function to push request content.
Endpoint.addStream and Client.initiateStream require the app layer
to provide an optional action for getting the next piece of content
for the request. This complicates life for the client if there is
a request body, since it must track the progress in the request.
It's simpler to give the client a more file-like interface, where
there is an action

  StreamContent -> IO ()

for sinking the content of a request in chunks. In this approach,
'addStream' has the type

  Connection
  -> StreamID
  -> Priority
  -> Bool
  -> IO (Maybe (StreamContent -> IO ()), IO StreamContent)

where the Bool indicates whether the stream is half-closed,
and the result includes an optional request 'pusher', and
a response 'puller'. The pusher is Nothing if the stream is
half-closed.

Similarly, 'initiateStream' has the result type

  IO (StreamID, Maybe (StreamContent -> IO ()), IO StreamContent)

and returns the stream ID, the optional request pusher, and
the response puller.

The StreamOptions type no longer has the 'streamOptsProducer'
field, but instead

  streamOptsHalfClosed :: Bool

which simply says whether the new stream should be started half-closed.

This simplifies 'addStream' slightly, since it no longer needs to
start a thread to pull request content from the app layer. The
interface for the app layer is also more intuitive. For a typical
request, it can now

 - initiate the stream
 - loop to push the request content
 - loop to pull the response content

For HTTP, this is probably typical use anyway, since the server
isn't likely to respond until it's read the full body of the
request.

The required modification in SGet was trivial, and it's still
able to make GET requests.