Add some Haddock docs to the Endpoint frame creation functions.
This improves the module documentation slightly, though the
synopsis in the generated HTML is still too wide because of the
very long type of 'synStreamFrame'. I'm reluctant to add a special
type for containing the various arguments to 'synStreamFrame' though,
so perhaps we'll just have to live with the ungainly synopsis.