Add FrameType and the WithFrameType and WithStream type classes.
One of things we need to do for stream error handling is send
RST_STREAM frames when we receive a frame for a stream that isn't
open. However, the spec says that we should not respond to a
RST_STREAM with another RST_STREAM, in order to prevent an
infinite loop of RST_STREAM frames.

That's a little inconvenient in the Endpoint input frame handlers,
because we're dealing the unboxed frames and so with several
distinct frame types. We could add a type class

 class IsRstStream a where
   isRstStream :: a -> Bool

but that seems awfully heavy-handed, particularly when we might
want to have a similar test for the other frame types.

Instead, I've added a FrameType enumeration, and a WithFrameType
class. Each of the specific frame types is in WithFrameType, along
with ControlFrame and Frame. With a very small amount of work,
the RawFrame type could be brought into the class too.

The WithFrameType class makes straightforward the function

  frameTypeName :: WithFrameType f => f -> String

I've also added the 'WithStream' type class, and instances for
the frame types that are associated with streams.

In the next commit, I'll use this stuff to slightly improve the
Endpoint input frame handlers.