Send RST_STREAM when we read a frame for a non-existent stream.
Now the Endpoint input frame handlers will send a RST_STREAM with
the InvalidStream status when we have no record of the stream.
In the process, I've exploited the new WithFrameType and WithStream
type classes to avoid hard-coding stream names, and to allow
passing the frames into 'forStream' and 'streamError'.

There are several things remaining to do here:

1. in addition to sending the RST_STREAM, we need to close the
   stream, at least from our side.

2. when the stream exists, but is supposed to be closed for the
   remote, we should instead use the StreamAlreadyClosed status

In order to do this additional handling, we'll need to change the
internals of Stream a little so it tracks which sides of the
stream are open. We'll then also be able to throw exceptions in the
request pusher and response puller as appropriate.

First, I'd like to change Stream so that its internals aren't
visible outside the Endpoint module. I'll do that in the next commit.