Prepare to move Endpoint.Stream into its own module, step 1.
In order to cleanly unit test the flow control provisions in streams,
I'd like to move the Stream data type into its own module. However,
at the moment the code for streams and connections is a little
entangled. Code in the Endpoint module, particularly in 'addStream'
and in the input frame handlers, feels free to reach into the
guts of a stream. It would be better to make Stream more
self-contained. In order to keep it self-contained, I'd like
to move it into its own module.

The focus of the change here is to extract a 'newStream' function
and rework 'addStream' to use it. 'newStream' concerns itself with
the stream internals, and 'addStream' concerns itself with the
interaction between the stream and the connection.

At first, I hit a problem where 'newStream' has to queue a
WINDOW_UPDATE frame on the enclosing connection's outgoing queue,
and therefore needs access to the connection. I realized that I
could undo this coupling by allowing StreamContent to carry
window updates. In order to allow a stream to reset itself, I've
also decided to add resets to StreamContent.

When building a new stream, one of the parameters is a sink for
outgoing content. Both the content pusher and the content puller
returned by 'newStream' use this sink action:

- the pusher uses it for the content it is giving, after waiting
  on the outgoing semaphore

- the puller uses it to send a window update after it has removed
  the next chunk of data from the incoming buffer

The only problem with this is that it allows the app layer to
deliberately send a window update or a reset. The app layer could
undermine the flow control algorithm by sending window updates,
or accidentally abort the stream for a bogus reason. It might be
a good idea to have two kinds of stream content, one for use in
the app layer, and the other for internal use with the stream's
outgoing sink. I'll consider that in the next couple of commits.

Well, I might have to rework the particulars, but the idea of
decoupling a stream and the enclosing connection is a good one.

There is a corresponding change in SGet to handle the incoming
window updates and resets (by ignoring them) but the change is minor.
SGet can still fetch the Google main page.

The next step will be to remove from Endpoint's input frame handlers
the direct references to the bounded buffer and the flow control
semaphore. Unless I decide to hide window updates and resets from
the app layer first.