Don't include window updates, resets in StreamContent after all.
Including window updates and resets allows the app layer to insert
a window update or an arbitrary reset in a stream which is a bad
idea. Instead, restore StreamContent to what it was before, and
add to NewStreamOpts functions for handling outgoing window
updates and stream resets. The pusher and puller returned by
'newStream' remains free of references to the enclosing connection.

This also undoes the previous changes to SGet.