Move Streams into their own module.
Move Stream and its closely-associated stuff out of the Endpoint
module and into the new Network.SPDY.Stream module. This completes
the separation of streams and the isolation of their internals
from their enclosing Connection.

There is some cleanup and documentation to do here, but then I
should be able to write some unit tests for streams. We'll also
need to track when the sides of a stream close, which will require
some internal rework.